+++ ninja-core/src/main/java/ninja/postoffice/PostofficeConstant.java
@@ -14,19 +14,23 @@
  * limitations under the License.
  */
 
-package ninja.migrations;
+package ninja.postoffice;
 
-public interface MigrationEngine {
+
+public interface PostofficeConstant {
     
-    /**
-     * Runs the migration scripts.
-     * 
-     * But only when ninja.migration.run=true
-     * 
-     * Implementations of migrate MUST BE safe in terms of multithreading and multiple instances 
-     * potentially running migrations at the same time...
-     * 
-     */
-    public void migrate();
+    public static final String postofficeImplementation = "postoffice.implementation";
+
+    public static final String smtpHost = "smtp.host";
+
+    public static final String smtpPort = "smtp.port";
+
+    public static final String smtpSsl = "smtp.ssl";
+
+    public static final String smtpUser = "smtp.user";
+
+    public static final String smtpPassword = "smtp.password";
+
+    public static final String smtpDebug = "smtp.debug";
 
 }